refactor: split invoke command facade internals
