+++ mhm/src/lib/command/idempotency.ts
@@ -1,0 +1,7 @@
+export function createIdempotencyKey(command: string): string {
+  const random =
+    typeof crypto !== "undefined" && "randomUUID" in crypto
+      ? crypto.randomUUID()
+      : `${Date.now()}-${Math.random().toString(16).slice(2)}`;
+  return `${command}:${random}`;
+}
