+++ mhm/src/lib/command/payload.ts
@@ -1,0 +1,6 @@
+export function buildInvokePayload(
+  args: Record<string, unknown> | undefined,
+  correlationId: string | undefined,
+): Record<string, unknown> | undefined {
+  return correlationId === undefined ? args : { ...(args ?? {}), correlationId };
+}
